@@ -30,9 +30,9 @@ class _FloatingBoxBottomSheet extends StatelessWidget {
   final SheetRequest request;
   final Function(SheetResponse) completer;
   const _FloatingBoxBottomSheet({
-    Key key,
-    this.request,
-    this.completer,
+    Key? key,
+    required this.request,
+    required this.completer,
   }) : super(key: key);
 
   @override
@@ -48,7 +48,7 @@ class _FloatingBoxBottomSheet extends StatelessWidget {
         mainAxisSize: MainAxisSize.min,
         children: [
           Text(
-            request.title,
+            request.title??'',
             style: TextStyle(
               fontSize: 20,
               fontWeight: FontWeight.bold,
@@ -57,7 +57,7 @@ class _FloatingBoxBottomSheet extends StatelessWidget {
           ),
           SizedBox(height: 10),
           Text(
-            request.description,
+            request.description??'',
             style: TextStyle(color: Colors.grey),
           ),
           Row(
@@ -67,14 +67,14 @@ class _FloatingBoxBottomSheet extends StatelessWidget {
               MaterialButton(
                 onPressed: () => completer(SheetResponse(confirmed: true)),
                 child: Text(
-                  request.secondaryButtonTitle,
+                  request.secondaryButtonTitle??'',
                   style: TextStyle(color: Theme.of(context).primaryColor),
                 ),
               ),
               TextButton(
                 onPressed: () => completer(SheetResponse(confirmed: true)),
                 child: Text(
-                  request.mainButtonTitle,
+                  request.mainButtonTitle??'',
                   style: TextStyle(color: Colors.white),
                 ),
                 style: ButtonStyle(
@@ -112,9 +112,9 @@ class GenericBottomSheet extends StatelessWidget {
   final Function(SheetResponse<GenericBottomSheetResponse>) completer;
 
   const GenericBottomSheet({
-    Key key,
-    this.request,
-    this.completer,
+    Key? key,
+    required this.request,
+    required this.completer,
   }) : super(key: key);
 
   @override
@@ -139,7 +139,7 @@ class GenericBottomSheet extends StatelessWidget {
           ),
           SizedBox(height: 10),
           Text(
-            request.description,
+            request.description??'',
             style: TextStyle(color: Colors.grey),
           ),
           Row(
@@ -152,7 +152,7 @@ class GenericBottomSheet extends StatelessWidget {
                   data: GenericBottomSheetResponse(message: 'SecondaryButton'),
                 )),
                 child: Text(
-                  request.secondaryButtonTitle,
+                  request.secondaryButtonTitle??'',
                   style: TextStyle(color: Theme.of(context).primaryColor),
                 ),
               ),
@@ -162,7 +162,7 @@ class GenericBottomSheet extends StatelessWidget {
                   data: GenericBottomSheetResponse(message: 'MainButton'),
                 )),
                 child: Text(
-                  request.mainButtonTitle,
+                  request.mainButtonTitle??'',
                   style: TextStyle(color: Colors.white),
                 ),
                 style: ButtonStyle(
