@@ -6,7 +6,7 @@ import 'package:stacked_services/stacked_services.dart';
 import '../../app/app.locator.dart';
 
 class SnackbarView extends StatelessWidget {
-  SnackbarView({Key key}) : super(key: key);
+  SnackbarView({Key? key}) : super(key: key);
 
   final _snackbarService = locator<SnackbarService>();
 
