@@ -28,4 +28,6 @@ import '../ui/views/second_screen.dart';
     ),
   ],
 )
-class App extends StackedApp {}
+class App {
+  /** This class has no puporse besides housing the annotation that generates the required functionality **/
+}
