@@ -6,12 +6,18 @@
 
 // ignore_for_file: public_member_api_docs
 
-import 'package:stacked/stacked.dart';
+import 'package:stacked_core/stacked_core.dart';
 import 'package:stacked_services/stacked_services.dart';
 
 final locator = StackedLocator.instance;
 
-void setupLocator() {
+Future<void> setupLocator(
+    {String? environment, EnvironmentFilter? environmentFilter}) async {
+// Register environments
+  locator.registerEnvironment(
+      environment: environment, environmentFilter: environmentFilter);
+
+// Register dependencies
   locator.registerLazySingleton(() => DialogService());
   locator.registerLazySingleton(() => SnackbarService());
   locator.registerLazySingleton(() => BottomSheetService());
